CoreNetworking:
Removed the -connectionWillPerformRequest: override method and simply put it in the -performWrite:forTag:withTimeout: method.
Updated the HTTP connection logic to read requests too for the server.
Completing the HTTP packet when reading a zero body request.
Keeping track of the transaction type, wether it started out with a request or one was created.